Tidying up test branch with some updates
